requirements.txt
alembic==1.7.3
autopep8==1.5.7
black==21.9b0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nltk==3.6.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dotenv==0.19.0
regex==2021.9.30
SQLAlchemy==1.4.25
textblob==0.15.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.3
typing-extensions==3.10.0.2
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.6.0
